realm: fix comparisons between signed and unsigned ints in procset module

+++ runtime/realm/procset/procset_module.cc
@@ -132,13 +132,13 @@ namespace Realm {
     {
       Module::create_processors(runtime);
       // for simplicity don't allow more that one procset per node for now
-      if (cfg_num_mp_procs > gasnet_nodes()) {
+      if (cfg_num_mp_procs > static_cast<int>(gasnet_nodes())) {
 	    log_procset.fatal() << "error num_mp_procs > number of nodes";
 	    assert(false);
       }
       if (cfg_num_mp_threads) {
         // if num_mp_procs is not set then assume one procset on every node
-        if (cfg_num_mp_procs == 0 || gasnet_mynode() < cfg_num_mp_procs ) { 
+        if (cfg_num_mp_procs == 0 || static_cast<int>(gasnet_mynode()) < cfg_num_mp_procs ) { 
           Processor p = runtime->next_local_processor_id();
           ProcessorImpl *pi = new LocalProcessorSet(p, runtime->core_reservation_set(),
             cfg_stack_size_in_mb << 20, cfg_num_mp_threads);
